+++ site/dat/docs/Changelog.md
@@ -12,6 +12,7 @@
  - recognize JMX variables in jmx2yaml
  - don't fail execution because of web browser
  - fix schedule size estimation and progress reporting in PBench
+ - fix PBench schedule reading crashes in Python3
 
 ## 1.4.3 <sup>14 apr 2016</sup>
  - bump up version for jmeter plugins installation to 1.4.0